soroban v0.0.1 Soroban.Email

A module for sending emails

Our emails can be sent as pure text, HTML or even as an application view, but to keep it simple we will use text only in the examples below. Check the Phoenix Framework docs (http://www.phoenixframework.org/docs/sending-email) for more details on how to send other types of emails

Functions

ask_confirm(email, link)

An email with a confirmation link in it.

ask_reset(email, link)

An with a link to reset the password.

invoice_html_email(email_address, invoice, jobs, total, company)

receipt_confirm(email)

An email acknowledging that the account has been successfully confirmed.

render(email, template, assigns \\ [])

Render an Phoenix template and set the body on the email

Pass an atom as the template name (:welcome_email) to render HTML and plain text emails. Use a string if you only want to render one type, e.g. “welcome_email.text” or “welcome_email.html”. Scroll to the top for more examples.