soroban v0.0.1 Soroban.Pdf

PDF helper functions

Functions

batch_zip(conn, invoicenum, clients)

Generate and zip PDF files

pdf_path()

Returns PDF path from config/config.exs

render(email, template, assigns \\ [])

Render an Phoenix template and set the body on the email

Pass an atom as the template name (:welcome_email) to render HTML and plain text emails. Use a string if you only want to render one type, e.g. “welcome_email.text” or “welcome_email.html”. Scroll to the top for more examples.

send_pdf(conn, id)

Build the filename from Invoice info and send the PDF to the client

send_zip(conn, invoicenum, filenames)

Sends the zipped PDFs to browser

to_html(invoice, jobs, total, company)

Generates HTML from the Email HTML template

Returns

to_pdf(invoice, jobs, total, company)

Generates PDF from HTML and saves to the filesystem if PDF file does not already exist in cache.