soroban v0.0.1 Soroban.Router.Helpers

Module with named helpers generated from Soroban.Router.

Functions

admin_path(conn_or_endpoint, action)

admin_path(conn_or_endpoint, action, params)

admin_url(conn_or_endpoint, action)

admin_url(conn_or_endpoint, action, params)

batch_path(conn_or_endpoint, action)

batch_path(conn_or_endpoint, action, params)

batch_path(conn_or_endpoint, action, id, params)

batch_url(conn_or_endpoint, action)

batch_url(conn_or_endpoint, action, params)

batch_url(conn_or_endpoint, action, id, params)

client_path(conn_or_endpoint, action)

client_path(conn_or_endpoint, action, params)

client_path(conn_or_endpoint, action, id, params)

client_url(conn_or_endpoint, action)

client_url(conn_or_endpoint, action, params)

client_url(conn_or_endpoint, action, id, params)

invoice_invoice_path(conn_or_endpoint, action, invoice_id)

invoice_invoice_path(conn_or_endpoint, action, invoice_id, params)

invoice_invoice_url(conn_or_endpoint, action, invoice_id)

invoice_invoice_url(conn_or_endpoint, action, invoice_id, params)

invoice_path(conn_or_endpoint, action)

invoice_path(conn_or_endpoint, action, params)

invoice_path(conn_or_endpoint, action, id, params)

invoice_url(conn_or_endpoint, action)

invoice_url(conn_or_endpoint, action, params)

invoice_url(conn_or_endpoint, action, id, params)

job_path(conn_or_endpoint, action)

job_path(conn_or_endpoint, action, params)

job_path(conn_or_endpoint, action, id, params)

job_url(conn_or_endpoint, action)

job_url(conn_or_endpoint, action, params)

job_url(conn_or_endpoint, action, id, params)

jobtype_path(conn_or_endpoint, action)

jobtype_path(conn_or_endpoint, action, params)

jobtype_path(conn_or_endpoint, action, id, params)

jobtype_url(conn_or_endpoint, action)

jobtype_url(conn_or_endpoint, action, params)

jobtype_url(conn_or_endpoint, action, id, params)

page_path(conn_or_endpoint, action)

page_path(conn_or_endpoint, action, params)

page_url(conn_or_endpoint, action)

page_url(conn_or_endpoint, action, params)

password_reset_path(conn_or_endpoint, action)

password_reset_path(conn_or_endpoint, action, params)

password_reset_url(conn_or_endpoint, action)

password_reset_url(conn_or_endpoint, action, params)

path(data, path)

Generates the path information including any necessary prefix.

service_path(conn_or_endpoint, action)

service_path(conn_or_endpoint, action, params)

service_path(conn_or_endpoint, action, id, params)

service_url(conn_or_endpoint, action)

service_url(conn_or_endpoint, action, params)

service_url(conn_or_endpoint, action, id, params)

session_path(conn_or_endpoint, action)

session_path(conn_or_endpoint, action, params)

session_path(conn_or_endpoint, action, id, params)

session_url(conn_or_endpoint, action)

session_url(conn_or_endpoint, action, params)

session_url(conn_or_endpoint, action, id, params)

setting_path(conn_or_endpoint, action)

setting_path(conn_or_endpoint, action, params)

setting_path(conn_or_endpoint, action, id, params)

setting_url(conn_or_endpoint, action)

setting_url(conn_or_endpoint, action, params)

setting_url(conn_or_endpoint, action, id, params)

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

url(data)

Generates the connection/endpoint base URL without any path information.

user_path(conn_or_endpoint, action)

user_path(conn_or_endpoint, action, params)

user_path(conn_or_endpoint, action, id, params)

user_url(conn_or_endpoint, action)

user_url(conn_or_endpoint, action, params)

user_url(conn_or_endpoint, action, id, params)